refactor: 중복 애니메이션/레이아웃을 ErrorLayout으로 추출하고 resetKeys 추가
- src/lib/animation.ts: easeOut, containerVariants, itemVariants 공용 프리셋 분리
- ErrorLayout 컴포넌트 신규 생성: NotFound·Error 페이지의 공통 motion wrapper 통합
- OverviewPlatformSection ErrorBoundary에 resetKeys={[roasRankingsData]} 추가
- OverviewBudgetSection ErrorBoundary에 resetKeys={[budget]} 추가
